Added the Circleizer
Using redundant objects to create one of the slowest image circleizers
out there.

Yes, BitmapImage used to be LabImage, and BitmapPixel used to be
LabPixel, and yes, this was originally not a circleizer, which is part
of why this program does things it doesn't have to do (as in, convert
every single pixel to a proprietary format within a proprietary format
(and then back again)) to get to the end result that it does.

"All can see the code from which I circle, but none can see the strategy
out of which circleization is evolved."