Deprecated the old, isolated the new.
The old Circleizer system was built on redundant systems on top of
redundant systems, notably the BitmapImage and BitmapPixel redundancies.

The new system isolates all the circleizing into the Circleizer class,
which makes the circleization faster and more readable.

The old classes have been deprecated and will probably be deleted soon.

...soon...