Added the color finder, deleted the deprecated.
ColorFinder uses a lengthy, unreadable, three-dimensional, and
orthogonal version of the shortest-splitline method to find distinct
colors with which to circleize with.

Also messed around with the Circleizer, switching on the exponential
circle sizes instead of the linearly-incrementing circle sizes. This
decreases circleization time for large images considerably (although it
still takes quite some time).